Swift Calc README.md
This program was my second iteration writing a calculator program. After writing the first iteration on my second day learning objective-C, I have revisted the app to rewrite it in swift, and showcase the skills and concepts I have learned since day two. I really focused on acheiving the cleanest code possible. I sought to make the view controller as 'dumb' as possible, so that its only task was simply to display data. In my Calculator class, my methods each accomplish one individual class. For instance to press a number key, I have broken that action into two helper methods, displayNumberButtonPresses and numberButtonPressed, to allow the boolean states to be updated and to display a number and perform a function in the correct order.